git-svn-id: http://svn.seqan.de/seqan/trunk@5081 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/library/seqan/index/index_qgram_openaddressing.h
@@ -171,7 +171,7 @@ namespace SEQAN_NAMESPACE_MAIN
 		typedef typename Host<TShape>::Type									TTextValue;
 		
 		unsigned long qgrams = _qgramQGramCount(index);
-		double max_qgrams = pow((double)ValueSize<TTextValue>::VALUE, weight(indexShape(index)));
+		double max_qgrams = pow((double)ValueSize<TTextValue>::VALUE, (double)weight(indexShape(index)));
 		if (max_qgrams <= qgrams)
 			qgrams = ceil(max_qgrams);
